:sparkles: LOVE-22-feat(handle-successful-session): User token is now a custom hook
